refactor: using long as parameter type

+++ oak-core/src/main/java/org/apache/jackrabbit/oak/plugins/index/statistics/Hash.java
@@ -32,7 +32,7 @@ public class Hash {
      * @param n the maximum of the result
      * @return the reduced value
      */
-    public static int reduce(int hash, int n) {
+    public static int reduce(long hash, int n) {
         // http://lemire.me/blog/2016/06/27/a-fast-alternative-to-the-modulo-reduction/
         return (int) (((hash & 0xffffffffL) * n) >>> 32);
     }
